lab7_bot/requirements.txt
pyTelegramBotAPI==4.14.0
psycopg2-binary==2.9.9
numpy==1.24.3
matplotlib==3.7.1
transformers==4.35.2
torch==2.1.1
pandas==2.0.3
SQLAlchemy==2.0.23
